fix(rtorrent): set default network.xmlrpc.size_limit to 8m
This resolves "XML-RPC request too large" errors when adding unusually large .torrent files.

The bugfix only applies to newly generated copies of rtorrent.rc, which are typically created the first time the Docker container is started. To apply this fix, you can either delete the existing rtorrent.rc file to allow the image to generate a new configuration with the fix included or manually edit the rtorrent.rc file to apply the changes yourself.

Fixes #30